+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [0.18.0] - Unreleased
 - Add support for checking invalid enum values and names. This introduces a new
   warning message "protobuf-enum-value"
+- Add support for type checking scalar fields. This introduces a new warning
+  message "protobuf-type-error"
 
 ## [0.17.0] - 2021-02-18
 - Add support for scalar repeated fields